+++ testdata/de/dm/intellij/liferay/language/java/LiferayJavaDeprecationInspectionTest/com/liferay/journal/model/JournalArticle.java
@@ -1,0 +1,4 @@
+package com.liferay.journal.model;
+
+public class JournalArticle {
+}